Crate rustc_front

Unstable (rustc_private)

The Rust compiler.

Note

This API is completely unstable and subject to change.

Modules

fold 	 [Unstable]
A Folder represents an HIR->HIR fold; it accepts a HIR piece, and returns a piece of the same type.

hir 	 [Unstable]

intravisit 	 [Unstable]
HIR walker. Each overridden visit method has full control over what happens with its node, it can do its own traversal of the node's children, call intravisit::walk_* to apply the default traversal algorithm, or prevent deeper traversal by doing nothing.

lowering 	 [Unstable]

print 	 [Unstable]

util 	 [Unstable]