Function rustc_front::attr::find_repr_attrs

pub fn find_repr_attrs(diagnostic: &SpanHandler, attr: &Attribute) -> Vec<ReprAttr>

Unstable (rustc_private)

Parse #[repr(...)] forms.

Valid repr contents: any of the primitive integral type names (see int_type_of_word, below) to specify enum discriminant type; C, to use the same discriminant size that the corresponding C enum would or C structure layout, and packed to remove padding.