Module rustc_front::hir

Unstable (rustc_private)

Reexports

pub use self::BindingMode::*;

pub use self::BinOp_::*;

pub use self::BlockCheckMode::*;

pub use self::CaptureClause::*;

pub use self::Decl_::*;

pub use self::ExplicitSelf_::*;

pub use self::Expr_::*;

pub use self::FunctionRetTy::*;

pub use self::ForeignItem_::*;

pub use self::ImplItem_::*;

pub use self::Item_::*;

pub use self::Mutability::*;

pub use self::Pat_::*;

pub use self::PathListItem_::*;

pub use self::PatWildKind::*;

pub use self::PrimTy::*;

pub use self::Stmt_::*;

pub use self::StructFieldKind::*;

pub use self::TraitItem_::*;

pub use self::Ty_::*;

pub use self::TyParamBound::*;

pub use self::UnOp::*;

pub use self::UnsafeSource::*;

pub use self::VariantKind::*;

pub use self::ViewPath_::*;

pub use self::Visibility::*;

pub use self::PathParameters::*;

Structs

AngleBracketedParameterData 	 [Unstable]
A path like Foo<'a, T>

Arg 	 [Unstable]
represents an argument in a function header

Arm 	 [Unstable]
represents one arm of a 'match'

BareFnTy 	 [Unstable]

Block 	 [Unstable]

Crate 	 [Unstable]

EnumDef 	 [Unstable]

Expr 	 [Unstable]
An expression

Field 	 [Unstable]

FieldPat 	 [Unstable]
A single field in a struct pattern

FnDecl 	 [Unstable]
Represents the header (not the body) of a function declaration

ForeignItem 	 [Unstable]

ForeignMod 	 [Unstable]

Generics 	 [Unstable]
Represents lifetimes and type parameters attached to a declaration of a function, enum, trait, etc.

ImplItem 	 [Unstable]

InlineAsm 	 [Unstable]

Item 	 [Unstable]
An item

Lifetime 	 [Unstable]

LifetimeDef 	 [Unstable]
A lifetime definition, eg 'a: 'b+'c+'d

Local 	 [Unstable]
Local represents a let statement, e.g., let <pat>:<ty> = <expr>;

MacroDef 	 [Unstable]
A macro definition, in this crate or imported from another.

MethodSig 	 [Unstable]
Represents a method's signature in a trait declaration, or in an implementation.

Mod 	 [Unstable]

MutTy 	 [Unstable]

ParenthesizedParameterData 	 [Unstable]
A path like Foo(A,B) -> C

Pat 	 [Unstable]

Path 	 [Unstable]
A "Path" is essentially Rust's notion of a name; for instance: std::cmp::PartialEq . It's represented as a sequence of identifiers, along with a bunch of supporting information.

PathSegment 	 [Unstable]
A segment of a path: an identifier, an optional lifetime, and a set of types.

PolyTraitRef 	 [Unstable]

QSelf 	 [Unstable]
The explicit Self type in a "qualified path". The actual path, including the trait and the associated item, is stored separately. position represents the index of the associated item qualified with this Self type.

StructDef 	 [Unstable]

StructField_ 	 [Unstable]

TraitItem 	 [Unstable]
Represents a method declaration in a trait declaration, possibly including a default implementation A trait method is either required (meaning it doesn't have an implementation, just a signature) or provided (meaning it has a default implementation).

TraitRef 	 [Unstable]
TraitRef's appear in impls.

Ty 	 [Unstable]

TyParam 	 [Unstable]

TypeBinding 	 [Unstable]

VariantArg 	 [Unstable]

Variant_ 	 [Unstable]

WhereBoundPredicate 	 [Unstable]
A type bound, eg for<'c> Foo: Send+Clone+'c

WhereClause 	 [Unstable]
A where clause in a definition

WhereEqPredicate 	 [Unstable]
An equality predicate (unsupported), e.g. T=int

WhereRegionPredicate 	 [Unstable]
A lifetime predicate, e.g. 'a: 'b+'c

Enums

BinOp_ 	 [Unstable]

BindingMode 	 [Unstable]

BlockCheckMode 	 [Unstable]

CaptureClause 	 [Unstable]

Constness 	 [Unstable]

Decl_ 	 [Unstable]

ExplicitSelf_ 	 [Unstable]
Represents the kind of 'self' associated with a method

Expr_ 	 [Unstable]

ForeignItem_ 	 [Unstable]
An item within an extern block

FunctionRetTy 	 [Unstable]

ImplItem_ 	 [Unstable]

ImplPolarity 	 [Unstable]

Item_ 	 [Unstable]

MatchSource 	 [Unstable]

Mutability 	 [Unstable]

PatWildKind 	 [Unstable]

Pat_ 	 [Unstable]

PathListItem_ 	 [Unstable]

PathParameters 	 [Unstable]

PrimTy 	 [Unstable]
Not represented directly in the AST, referred to by name through a ty_path.

Stmt_ 	 [Unstable]

StructFieldKind 	 [Unstable]

TraitBoundModifier 	 [Unstable]
A modifier on a bound, currently this is only used for ?Sized, where the modifier is Maybe. Negative bounds should also be handled here.

TraitItem_ 	 [Unstable]

TyParamBound 	 [Unstable]
The AST represents all type param bounds as types. typeck::collect::compute_bounds matches these against the "special" built-in traits (see middle::lang_items) and detects Copy, Send and Sync.

Ty_ 	 [Unstable]
The different kinds of types recognized by the compiler

UnOp 	 [Unstable]

UnsafeSource 	 [Unstable]

Unsafety 	 [Unstable]

VariantKind 	 [Unstable]

ViewPath_ 	 [Unstable]

Visibility 	 [Unstable]

WherePredicate 	 [Unstable]
A single predicate in a where clause

Type Definitions

BinOp 	 [Unstable]

Decl 	 [Unstable]

ExplicitSelf 	 [Unstable]

PathListItem 	 [Unstable]

Stmt 	 [Unstable]
A statement

StructField 	 [Unstable]

TyParamBounds 	 [Unstable]

Variant 	 [Unstable]

ViewPath 	 [Unstable]