Enum rustc_front::hir::Expr_

pub enum Expr_ {
    ExprBox(P<Expr>),
    ExprVec(HirVec<P<Expr>>),
    ExprCall(P<Expr>, HirVec<P<Expr>>),
    ExprMethodCall(Spanned<Name>, HirVec<P<Ty>>, HirVec<P<Expr>>),
    ExprTup(HirVec<P<Expr>>),
    ExprBinary(BinOp, P<Expr>, P<Expr>),
    ExprUnary(UnOp, P<Expr>),
    ExprLit(P<Lit>),
    ExprCast(P<Expr>, P<Ty>),
    ExprType(P<Expr>, P<Ty>),
    ExprIf(P<Expr>, P<Block>, Option<P<Expr>>),
    ExprWhile(P<Expr>, P<Block>, Option<Ident>),
    ExprLoop(P<Block>, Option<Ident>),
    ExprMatch(P<Expr>, HirVec<Arm>, MatchSource),
    ExprClosure(CaptureClause, P<FnDecl>, P<Block>),
    ExprBlock(P<Block>),
    ExprAssign(P<Expr>, P<Expr>),
    ExprAssignOp(BinOp, P<Expr>, P<Expr>),
    ExprField(P<Expr>, Spanned<Name>),
    ExprTupField(P<Expr>, Spanned<usize>),
    ExprIndex(P<Expr>, P<Expr>),
    ExprPath(Option<QSelf>, Path),
    ExprAddrOf(Mutability, P<Expr>),
    ExprBreak(Option<Spanned<Ident>>),
    ExprAgain(Option<Spanned<Ident>>),
    ExprRet(Option<P<Expr>>),
    ExprInlineAsm(InlineAsm),
    ExprStruct(Path, HirVec<Field>, Option<P<Expr>>),
    ExprRepeat(P<Expr>, P<Expr>),
}

Unstable (rustc_private)

Variants

ExprBox
Unstable (rustc_private)

A box x expression.

ExprVec
Unstable (rustc_private)

An array ([a, b, c, d])

ExprCall
Unstable (rustc_private)

A function call

The first field resolves to the function itself, and the second field is the list of arguments

ExprMethodCall
Unstable (rustc_private)

A method call (x.foo::<Bar, Baz>(a, b, c, d))

The Spanned<Name> is the identifier for the method name. The vector of Tys are the ascripted type parameters for the method (within the angle brackets).

The first element of the vector of Exprs is the expression that evaluates to the object on which the method is being called on (the receiver), and the remaining elements are the rest of the arguments.

Thus, x.foo::<Bar, Baz>(a, b, c, d) is represented as ExprMethodCall(foo, [Bar, Baz], [x, a, b, c, d]).

ExprTup
Unstable (rustc_private)

A tuple ((a, b, c ,d))

ExprBinary
Unstable (rustc_private)

A binary operation (For example: a + b, a * b)

ExprUnary
Unstable (rustc_private)

A unary operation (For example: !x, *x)

ExprLit
Unstable (rustc_private)

A literal (For example: 1u8, "foo")

ExprCast
Unstable (rustc_private)

A cast (foo as f64)

ExprType
Unstable (rustc_private)

ExprIf
Unstable (rustc_private)

An if block, with an optional else block

if expr { block } else { expr }

ExprWhile
Unstable (rustc_private)

A while loop, with an optional label

'label: while expr { block }

ExprLoop
Unstable (rustc_private)

Conditionless loop (can be exited with break, continue, or return)

'label: loop { block }

ExprMatch
Unstable (rustc_private)

A match block, with a source that indicates whether or not it is the result of a desugaring, and if so, which kind.

ExprClosure
Unstable (rustc_private)

A closure (for example, move |a, b, c| {a + b + c})

ExprBlock
Unstable (rustc_private)

A block ({ ... })

ExprAssign
Unstable (rustc_private)

An assignment (a = foo())

ExprAssignOp
Unstable (rustc_private)

An assignment with an operator

For example, a += 1.

ExprField
Unstable (rustc_private)

Access of a named struct field (obj.foo)

ExprTupField
Unstable (rustc_private)

Access of an unnamed field of a struct or tuple-struct

For example, foo.0.

ExprIndex
Unstable (rustc_private)

An indexing operation (foo[2])

ExprPath
Unstable (rustc_private)

Variable reference, possibly containing :: and/or type parameters, e.g. foo::bar::.

Optionally "qualified", e.g. <HirVec<T> as SomeTrait>::SomeType.

ExprAddrOf
Unstable (rustc_private)

A referencing operation (&a or &mut a)

ExprBreak
Unstable (rustc_private)

A break, with an optional label to break

ExprAgain
Unstable (rustc_private)

A continue, with an optional label

ExprRet
Unstable (rustc_private)

A return, with an optional value to be returned

ExprInlineAsm
Unstable (rustc_private)

Output of the asm!() macro

ExprStruct
Unstable (rustc_private)

A struct literal expression.

For example, Foo {x: 1, y: 2}, or Foo {x: 1, .. base}, where base is the Option<Expr>.

ExprRepeat
Unstable (rustc_private)

A vector literal constructed from one repeated element.

For example, [1u8; 5]. The first expression is the element to be repeated; the second is the number of times to repeat it.