Module rustc_front::util

Unstable (rustc_private)

Structs

IdVisitor 	 [Unstable]

Functions

binop_to_string 	 [Unstable]

compute_id_range_for_fn_body 	 [Unstable]
Computes the id range for a single fn body, ignoring nested items.

empty_generics 	 [Unstable]

ident_to_path 	 [Unstable]

is_by_value_binop 	 [Unstable]
Returns true if the binary operator takes its arguments by value

is_by_value_unop 	 [Unstable]
Returns true if the unary operator takes its argument by value

is_comparison_binop 	 [Unstable]

is_path 	 [Unstable]

is_shift_binop 	 [Unstable]

lazy_binop 	 [Unstable]

stmt_id 	 [Unstable]

unop_to_string 	 [Unstable]

walk_pat 	 [Unstable]